+++ src/systems/physics/Physics.cc
@@ -973,6 +973,8 @@ void PhysicsPrivate::UpdatePhysics(EntityComponentManager &_ecm)
         if (jointIt == this->entityJointMap.end())
           return true;
 
+        auto force = _ecm.Component<components::JointForceCmd>(_entity);
+
         // Model is out of battery
         if (this->entityOffMap[_ecm.ParentEntity(_entity)])
         {
@@ -980,15 +982,6 @@ void PhysicsPrivate::UpdatePhysics(EntityComponentManager &_ecm)
           for (std::size_t i = 0; i < nDofs; ++i)
           {
             jointIt->second->SetForce(i, 0);
-
-            // TODO(anyone): Only for diff drive, which does not use
-            //   JointForceCmd. Remove when it does.
-            auto jointVelFeature = entityCast(_entity, jointIt->second,
-                this->entityJointVelocityCommandMap);
-            if (jointVelFeature)
-            {
-              jointVelFeature->SetVelocityCommand(i, 0);
-            }
           }
           return true;
         }
@@ -1046,7 +1039,6 @@ void PhysicsPrivate::UpdatePhysics(EntityComponentManager &_ecm)
             }
         }
 
-        auto force = _ecm.Component<components::JointForceCmd>(_entity);
         auto velCmd = _ecm.Component<components::JointVelocityCmd>(_entity);
 
         if (force)