+++ Changelog.md
@@ -2,6 +2,9 @@
 
 ### Ignition Gazebo 2.xx.xx (20XX-XX-XX)
 
+1. Allow battery plugin to work with joint force systems.
+    * [Pull Request 120](https://github.com/ignitionrobotics/ign-gazebo/pull/120)
+
 1. Make breadcrumb static after specified time
     * [Pull Request 90](https://github.com/ignitionrobotics/ign-gazebo/pull/90)